FAC-416: Repository refactored to hold and use UpdateRole
