+++ flickit-assessment-common/src/main/java/org/flickit/assessment/common/application/domain/assessment/AssessmentPermission.java
@@ -26,5 +26,6 @@ public enum AssessmentPermission {
     VIEW_SUBJECT_PROGRESS,
     VIEW_SUBJECT_REPORT,
     CREATE_ADVICE,
-    GRANT_USER_ASSESSMENT_ROLE
+    GRANT_USER_ASSESSMENT_ROLE,
+    UPDATE_USER_ASSESSMENT_ROLE
 }
