+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/domain/AssessmentUserRole.java
@@ -83,7 +83,8 @@ public enum AssessmentUserRole {
             CREATE_ASSESSMENT,
             DELETE_ASSESSMENT,
             UPDATE_ASSESSMENT,
-            GRANT_USER_ASSESSMENT_ROLE));
+            GRANT_USER_ASSESSMENT_ROLE,
+            UPDATE_USER_ASSESSMENT_ROLE));
 
         private final Set<AssessmentPermission> permissions;
     }
